--- a/reclass/__init__.py
+++ b/reclass/__init__.py
@@ -19,7 +19,7 @@ def get_storage(storage_type, nodes_uri, classes_uri, compose_node_name, **kwarg
     storage_class = StorageBackendLoader(storage_type).load()
     return MemcacheProxy(storage_class(nodes_uri, classes_uri, compose_node_name, **kwargs))
 
-def get_path_mangler(storage_type,**kwargs):
+def get_path_mangler(storage_type, **kwargs):
     return StorageBackendLoader(storage_type).path_mangler()
 
 def output(data, fmt, pretty_print=False, no_refs=False):