… reclass/values/tests/test_compitem.py
… reclass/values/tests/test_item.py
… reclass/values/tests/test_listitem.py
… reclass/values/tests/test_refitem.py
… reclass/values/tests/test_scaitem.py
… reclass/values/valuelist.py
… reclass/version.py