Tests added, code refactored some more. Parsers are updated and cleaned up.