@AndrewPickford 2018-08-28
The non parser related changes should be fine. The parser usage changes in reclass/values/parser.py and the knock on changes to  reclass/reclass/values/parser_funcs.py have a problem.

Changing to use a single parser by default in reclass/values/parser.py lines 32 to 35 will incur a large speed penalty. The full parser (the object returned by  get_ref_parser in the non refactored code) is very slow. This is the reason for using the $ test in the non refactored Parse class parse method and for having a second simplified parser that only deals with strings containing a single reference. The slow down is fine for testing but for nodes with large numbers of complex classes it's bad and for nodes with inv queries the slow down is even worse.

For this refactoring I would not merge the parser changes. Instead split them off and tackle them in smaller chunks - separate the parser changes into the simple format changes (like removing underscores) and then the logic changes to the parser. I'd advise this for both the parser and the interpolation done in the Parameters/Exports classes as they are both nastily complex areas of the reclass code.