@epcim 2018-09-01
I agree, let's split it into multiple patches and for the patch to parser check performance. Thanks