@epcim 2018-10-12
Alex, any idea when you will have time to split to "syntax" update and the part that touch the parser only.
I would like to avoid conflicts, as they will always appear as we add new features.