@a-ovchinnikov 2018-10-18
Petr, I think, I will have time to address comments and suggestions in the next few days. 

@AndrewPickford - could you please share data which shows that regular parser incurs large speed penalty? My tests show that it is just about two times slower than simple one, but I have to admit that I benchmarked them on synthetic input only.