@AndrewPickford 2018-10-18
@a-ovchinnikov Checked the timing with my current setup. A factor of a little over 2 looks to be about right between the current code and just using the full parser. I thought it was higher, but your synthetic benchmark matches my setup timings for this.

That said for my slowest nodes reclass takes 11 seconds currently and increases to 24 seconds just using the full parser. With about a 10% measurement uncertainty on those numbers. The more simple node times go from 3-4 to 7-9 seconds. I think these times will be at the high end for reclass runs - I'm using git and even my simple nodes still have a fair bit of reclass config.

In terms of salt runs these times are a non issue. Doing a full state run over even one of my simple nodes takes around 30 to 60 seconds, so an additional few seconds from reclass would not be a problem. But running reclass on the command line should be as fast as possible, waiting those extra 3 to 12 seconds adds up.

Also the original reclass is a bit faster still, at least on inputs that both can interpret. Minimising any speed regression between the two was the initial reason for the speed ups and the simplified parser.