@a-ovchinnikov 2018-10-19
I separated general refactoring from parser changes.