@AndrewPickford 2018-10-23
Ran some tests on with my setup. The refactored code produced identical output to the current develop branch so all good there. However the slow down has gotten even worse for my more complex nodes. Simple nodes again take about twice as long, but the more complex nodes take about 5 times as long. Going from 3.5 to 17.9 seconds in one case and for the worst case from 10.9 to 52.9 seconds.

The tests we currently have aren't any good for timing tests, run_tests.py on my laptop takes about 2.8 secs for the refactored code and 2.2 secs for the develop branch. And there's at least a +/- 0.1 sec uncertainty on these numbers. If would help to have a speed test, but it's a fair bit of work to make something large and realistic enough without copying a real reclass setup which will usually have non public information in it somewhere.