@a-ovchinnikov 2018-10-24
The slowdown of run_tests.py is expected: it is a runner for unit tests, I am adding a dozen or so of new unit tests some of which take time to execute, thus in the end we get longer execution times for the runner.

However the slowdown for complex nodes is rather surprising given that this PR consists of mostly minor deduplication tweaks. I definitely need to run few more tests. Meanwhile could you please provide some details regarding how you measure this slowdown? Are the values you shared mean? What about deviations? What was the number of tests? Also could you please include hash of a commit for which you do the measurements, just to play it the safest way  and ensure we are both measurng the same thing?