@AndrewPickford 2018-10-26
@a-ovchinnikov  The issue with run_test.py for timing tests is that it isn't doing enough work for the timing to be meaning full. And as you say adding a more tests (a good thing!) slows things down anyway.

I'm doing simple timing tests just running reclass with time, so:

time reclass --nodeinfo node1 -r

Unless I get a strange or long result I'm only running a test once. I'm not doing anything statistical here - I'm just testing for large changes. From running these tests few times the lowest uncertainty on these sorts of simple tests is around 0.1 secs, for the short few second tests. Longer tests have more uncertainty but I don't have any firm numbers for this.

The results from a few days ago were done against current salt-formulas:develop branch (d1a099ab91a6d6feec7031a0d1aa16427da2d6d5) and your a-ovchinnikov:develop branch (bb76a54ff6aa3ad024ff8e8d21185ffe59911d97).

Looking at the changes to reclass/values/parser.py, you're calling parsers.get_ref_parser and parsers.get_simple_ref_parser on each call to the Parse.parse method. That's probably where most of the slow down is coming from.