@epcim 2018-10-31
I am about to merge it, @AndrewPickford have you tested with the latest version? May you update your review. Thanks