@AndrewPickford 2018-11-01
Looks good now. Speed issue is fixed.